story: password 스토리북 기능 등록
